mt8173: Add board directory to hold board specific code
We add a board directory and board parameters passing mechanism to
separate board specific code from soc. The board related parameters can
be passed from earlier boot stage to ARM TF during runtime. We define
two types of board parameters in this patch, external buck type for
suspending soc cluster 1, and the power notify gpio for powering off
the system. We pass the board parameters to ARM TF via shared memory.
MT8173 platform code can know the shared memory address from
bl31_plat_params structure, i.e., we only deal with the !RESET_TO_BL31
case in this patch.

Besides, memory layout could be different for different boards or
project. And, we may not want to include all the board related drivers
in BL31 binary. Hence, we add some build flags to determine which
memory layout to use and which board drivers to be included.

+++ plat/mediatek/soc/mt8173/drivers/pmic/pmic_wrap_init.c
@@ -118,8 +118,8 @@ static inline uint32_t wait_for_state_ready(uint32_t timeout_us,
 
 static int32_t pwrap_wacs2(uint32_t write,
 		    uint32_t adr,
-		    uint32_t wdata,
-		    uint32_t *rdata,
+		    uint16_t wdata,
+		    uint16_t *rdata,
 		    uint32_t init_check)
 {
 	uint32_t reg_rdata = 0;
@@ -176,12 +176,12 @@ FAIL:
 
 /* external API for pmic_wrap user */
 
-int32_t pwrap_read(uint32_t adr, uint32_t *rdata)
+int32_t pwrap_read(uint32_t adr, uint16_t *rdata)
 {
 	return pwrap_wacs2(0, adr, 0, rdata, 1);
 }
 
-int32_t pwrap_write(uint32_t adr, uint32_t wdata)
+int32_t pwrap_write(uint32_t adr, uint16_t wdata)
 {
 	return pwrap_wacs2(1, adr, wdata, 0, 1);
 }
